ref(auto-install): Extract OTel version checker
Move OpenTelemetry downgrade detection and message formatting out of the Gradle task so the task only handles wiring and execution.

Add focused unit coverage for artifact detection and failure message formatting.
